LayoutTests/http/tests/security/xssAuditor/full-block-link-onclick-expected.txt

CONSOLE MESSAGE: line 4: Refused to execute a JavaScript script. Source code of script found within request.

There should be no content in the iframe below:



--------
Frame: 'frame'
--------
